+++ lib/yt.rb
@@ -7,6 +7,8 @@ require 'yt/models/match_policy'
 require 'yt/models/playlist'
 require 'yt/models/playlist_item'
 require 'yt/models/video'
+require 'yt/models/ownership'
+require 'yt/models/video_advertising_options'
 
 # An object-oriented Ruby client for YouTube.
 # Helps creating applications that need to interact with YouTube objects.