+++ lib/yt/collections/base.rb
@@ -46,6 +46,16 @@ module Yt
       def apply_where_params!(params = {})
         params.merge!(@where_params ||= {})
       end
+
+      # If we dropped support for ActiveSupport 3, then we could simply
+      # invoke transform_keys!{|key| key.to_s.underscore.to_sym}
+      def underscore_keys!(hash)
+        hash.dup.each_key{|key| hash[underscore key] = hash.delete key}
+      end
+
+      def underscore(value)
+        value.to_s.underscore.to_sym
+      end
     end
   end
 end